:wrench: (frontend) refactored analysis components
