+++ frontend/src/app/dashboard/page.tsx
@@ -2,8 +2,7 @@
 
 import React from 'react';
 import ProtectedRoute from '@/components/ProtectedRoute';
-//import ExcelAnalysis from '@/components/dashboard/ExcelAnalysis';
-import AnalysisComponents from '@/components/dashboard/TableAnalysis/AnalysisComponents';
+import AnalysisComponents from '@/components/dashboard/TableAnalysis/Main';
 import Navbar from '@/components/navbar/Navbar';
 
 const ExcelAnalyzer = () => {
