+++ frontend/src/components/dashboard/TableAnalysis/Home/LoadingOverlay.tsx
@@ -1,0 +1,28 @@
+import React from 'react';
+import { Spin } from 'antd';
+
+interface LoadingOverlayProps {
+  isLoading: boolean;
+}
+
+const LoadingOverlay: React.FC<LoadingOverlayProps> = ({ isLoading }) => {
+  if (!isLoading) return null;
+  return (
+    <div style={{
+      position: 'absolute',
+      top: 0,
+      left: 0,
+      width: '100%',
+      height: '100%',
+      backgroundColor: 'rgba(255,255,255,0.8)',
+      zIndex: 9999,
+      display: 'flex',
+      justifyContent: 'center',
+      alignItems: 'center',
+    }}>
+      <Spin size="large" />
+    </div>
+  );
+};
+
+export default LoadingOverlay;
